CLOUDP-64640: moved wait_for_condition to a new file, refactored e2e script to use it
